refactor : 상수 파일로 타입 및 인터페이스 이동
